+++ src/starkinfra/issuing_token_design.clj
@@ -85,7 +85,7 @@
     - `user` [map, default nil]: Project or Organization map returned from starkinfra.user/project or starkinfra.user/organization. Only necessary if starkinfra.settings/user has not been set.
 
   ## Return:
-    - IssuingTokenDesign pdf file content"
+    - IssuingTokenDesign pdf file content, as a byte array"
   ([id]
    (get-content @credentials (resource) id "pdf" {}))